server/nexus/sandbox/docker/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pyautogui==0.9.54
pillow==10.2.0
pydantic==2.6.1
pytesseract==0.3.13

# Nexus Python Tools Dependencies
tavily-python==0.3.0
httpx==0.25.2
python-dotenv==1.0.0
lxml==4.9.3
beautifulsoup4==4.12.2

playwright
requests
pandas
numpy
matplotlib
seaborn
yfinance
plotly
scipy
scikit-learn
jupyterlab